Revert "SDMEXT-616: Implementing user token flow;Ccahing of user token: basic create attachment flow"
