+++ sdm/src/main/java/com/sap/cds/handler/SDMAttachmentsHandler.java
@@ -1,6 +1,5 @@
-package com.sap.cds.sdm.service.handler;
+package com.sap.cds.handler;
 
-import com.sap.cds.feature.attachments.generated.cds4j.sap.attachments.Attachments;
 import com.sap.cds.feature.attachments.service.AttachmentService;
 import com.sap.cds.feature.attachments.service.model.servicehandler.AttachmentCreateEventContext;
 import com.sap.cds.feature.attachments.service.model.servicehandler.AttachmentMarkAsDeletedEventContext;
@@ -8,18 +7,19 @@ import com.sap.cds.feature.attachments.service.model.servicehandler.AttachmentRe
 import com.sap.cds.feature.attachments.service.model.servicehandler.AttachmentRestoreEventContext;
 import com.sap.cds.services.handler.EventHandler;
 import com.sap.cds.services.handler.annotations.On;
-import com.sap.cds.services.handler.annotations.ServiceName;
-import java.io.IOException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-@ServiceName(value = "*", type = AttachmentService.class)
-public class SDMAttachmentsServiceHandler implements EventHandler {
+public class SDMAttachmentsHandler implements EventHandler {
+  private static final Logger logger = LoggerFactory.getLogger(SDMAttachmentsHandler.class);
 
   @On(event = AttachmentService.EVENT_CREATE_ATTACHMENT)
-  public void createAttachment(AttachmentCreateEventContext context) throws IOException {
-    var contentId = (String) context.getAttachmentIds().get(Attachments.ID);
-    context.setIsInternalStored(true);
-    context.setContentId(contentId);
-    context.setCompleted();
+  public void createAttachment(AttachmentCreateEventContext context) {
+    logger.info(
+        "Default Attachment Service handler called for creating attachment for entity name: {}",
+        context.getAttachmentEntity().getQualifiedName());
+    logger.info("Other details: {}", context.getAttachmentIds());
+    logger.info("Other details entity: {}", context.getAttachmentEntity());
   }
 
   @On(event = AttachmentService.EVENT_MARK_ATTACHMENT_AS_DELETED)
