+++ sdm/src/main/java/com/sap/cds/model/SDMCredentials.java
@@ -1,7 +1,6 @@
-package com.sap.cds.sdm.model;
+package com.sap.cds.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import java.io.InputStream;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,8 +11,13 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class CmisDocument {
-  private String fileName;
-  private InputStream content;
-  private String parentId;
+public class SDMCredentials {
+
+  private String url;
+
+  private String baseTokenUrl;
+
+  private String clientId;
+
+  private String clientSecret;
 }
