--- sdm/src/main/java/com/sap/cds/sdm/caching/CacheKey.java
@@ -1,13 +1,0 @@
-package com.sap.cds.sdm.caching;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CacheKey {
-  private String email;
-  private String expiration;
-}
