--- sdm/src/main/java/com/sap/cds/sdm/model/SDMCredentials.java
@@ -1,23 +1,0 @@
-package com.sap.cds.sdm.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class SDMCredentials {
-
-  private String url;
-
-  private String baseTokenUrl;
-
-  private String clientId;
-
-  private String clientSecret;
-}
