--- sdm/src/main/java/com/sap/cds/sdm/persistence/DBQuery.java
@@ -1,3 +1,0 @@
-package com.sap.cds.sdm.persistence;
-
-public class DBQuery {}
