--- sdm/src/main/java/com/sap/cds/sdm/service/RegisterService.java
@@ -1,7 +1,0 @@
-package com.sap.cds.sdm.service;
-
-import com.sap.cds.services.Service;
-
-public interface RegisterService extends Service {
-  String SDM_NAME = "SDMAttachmentService$Default";
-}
