--- sdm/src/main/java/com/sap/cds/sdm/util/HelperUtil.java
@@ -1,3 +1,0 @@
-package com.sap.cds.sdm.util;
-
-public class HelperUtil {}
