--- sdm/src/main/resources/META-INF/services/com.sap.cds.services.runtime.CdsRuntimeConfiguration
@@ -1,1 +1,0 @@
-com.sap.cds.sdm.configuration.Registration
