+++ sdm/src/main/resources/cds/com.sap.cds/sdm/attachments-annotations.cds
@@ -3,10 +3,21 @@ using {
     sap.attachments.Attachments
 } from './attachments';
 
-annotate MediaData with @UI.MediaResource: { Stream: content } {
-  content  @Core.MediaType: mimeType @odata.draft.skip;
-  mimeType @Core.IsMediaType;
-  status @readonly;
+annotate MediaData with @UI.MediaResource: {Stream: content} {
+    content   @(
+        title                      : '{i18n>attachment_content}',
+        Core.MediaType             : mimeType,
+        ContentDisposition.Filename: fileName,
+        ContentDisposition.Type    : 'inline'
+    );
+    mimeType  @(
+        title: '{i18n>attachment_mimeType}',
+        Core.IsMediaType
+    );
+    fileName  @(title: '{i18n>attachment_fileName}');
+    status    @(title: '{i18n>attachment_status}');
+    contentId @(UI.Hidden: true);
+    scannedAt @(UI.Hidden: true);
 }
 
 annotate Attachments with @UI: {
@@ -22,12 +33,11 @@ annotate Attachments with @UI: {
         {Value: createdBy},
         {Value: note}
     ]
+} {
+    note       @(title: '{i18n>attachment_note}');
+    modifiedAt @(odata.etag);
 }
-{
-  content
-    @Core.ContentDisposition: { Filename: fileName }
-    @Core.Immutable
-}
+
 annotate Attachments with @Common: {SideEffects #ContentChanged: {
     SourceProperties: [content],
     TargetProperties: ['status']
