+++ sdm/src/main/resources/cds/com.sap.cds/sdm/attachments.cds
@@ -24,5 +24,4 @@ aspect MediaData           @(_is_media_data) {
 
 aspect Attachments : cuid, managed, MediaData {
     note : String;
-
 }
